Ajout de ChocoSolver et génération des premiers planning

+++ src/main/java/solver/test.java
@@ -324,7 +324,7 @@ public class test {
                 Arrays.asList(
                         new Contrainte(
 
-                                Arrays.asList(new Integer(1),new Integer(2)),
+                                Arrays.asList(new Integer(2), new Integer(1)),
                                 1500,
                                 3000,
                                 20,
